@@ -1,9 +1,10 @@
 [package]
 name = "conlang_fmt"
-version = "1.1.0"
+version = "1.2.0"
 authors = ["Emily Crandall Fleischman <emilycf@mit.edu>"]
-edition = "2018"
+edition = "2021"
 
 [dependencies]
-failure = "0.1.6"
-itertools = "0.8.1"
+anyhow = "1.0.79"
+itertools = "0.12"
+thiserror = "1.0.56"
