@@ -147,9 +147,10 @@ impl GlossLine {
     }
 }
 
-#[derive(Debug, Eq, PartialEq)]
+#[derive(Debug, Default, Eq, PartialEq)]
 pub enum GlossLineType {
     NoSplit,
+    #[default]
     Split,
 }
 
@@ -166,11 +167,5 @@ impl GlossLineType {
                 _ => Some(param),
             },
         })
-    }
-}
-
-impl Default for GlossLineType {
-    fn default() -> GlossLineType {
-        GlossLineType::Split
     }
 }
